Fixing rounding after conversions.

+++ cryptocurrency/conversion_table.py
@@ -97,19 +97,23 @@ def get_conversion_table(client, exchange_info, offset_s=0, as_pair=False):
     conversion_table['USDT_open'] = \
         conversion_table.apply(lambda x: convert_price(size=1, from_asset=x['base_asset'], to_asset='USDT', 
                                                        conversion_table=conversion_table, 
-                                                       exchange_info=exchange_info, key='open'), axis='columns')
+                                                       exchange_info=exchange_info, key='open'), 
+                               axis='columns').astype(float)
     conversion_table['USDT_high'] = \
         conversion_table.apply(lambda x: convert_price(size=1, from_asset=x['base_asset'], to_asset='USDT', 
                                                        conversion_table=conversion_table, 
-                                                       exchange_info=exchange_info, key='close'), axis='columns')
+                                                       exchange_info=exchange_info, key='close'), 
+                               axis='columns').astype(float)
     conversion_table['USDT_low'] = \
         conversion_table.apply(lambda x: convert_price(size=1, from_asset=x['base_asset'], to_asset='USDT', 
                                                        conversion_table=conversion_table, 
-                                                       exchange_info=exchange_info, key='close'), axis='columns')
+                                                       exchange_info=exchange_info, key='close'), 
+                               axis='columns').astype(float)
     conversion_table['USDT_price'] = \
         conversion_table.apply(lambda x: convert_price(size=1, from_asset=x['base_asset'], to_asset='USDT', 
                                                        conversion_table=conversion_table, 
-                                                       exchange_info=exchange_info), axis='columns')
+                                                       exchange_info=exchange_info), 
+                               axis='columns').astype(float)
     conversion_table['rolling_USDT_base_volume'] = \
         conversion_table['rolling_base_volume'] * conversion_table['USDT_price']
     conversion_table['rolling_USDT_quote_volume'] = \
@@ -119,12 +123,12 @@ def get_conversion_table(client, exchange_info, offset_s=0, as_pair=False):
         conversion_table.apply(lambda x: convert_price(size=x['bid_price'], from_asset=x['base_asset'], 
                                                        to_asset='USDT', conversion_table=conversion_table, 
                                                        exchange_info=exchange_info), 
-                               axis='columns')
+                               axis='columns').astype(float)
     conversion_table['USDT_ask_price'] = \
         conversion_table.apply(lambda x: convert_price(size=x['ask_price'], from_asset=x['base_asset'], 
                                                        to_asset='USDT', conversion_table=conversion_table, 
                                                        exchange_info=exchange_info), 
-                               axis='columns')
+                               axis='columns').astype(float)
     conversion_table['USDT_bid_volume'] = \
         conversion_table['bid_volume'] * conversion_table['USDT_bid_price']
     conversion_table['USDT_ask_volume'] = \
